requirements.txt
# Core Dependencies
upstox-python-sdk==2.19.0
pandas>=2.3.3
numpy>=1.26.2
python-dotenv>=1.0.0

# Technical Analysis
# ta-lib==0.4.28  # Requires compilation, skipping for now
# pandas-ta>=0.3.14b0  # Requires numba which doesn't support Python 3.14 yet

# Data Visualization
matplotlib>=3.8.2
plotly>=5.18.0
seaborn>=0.13.0

# API & Web
requests>=2.31.0
websocket-client>=1.7.0

# Async Support
aiohttp>=3.9.1

# Database
sqlalchemy>=2.0.23

# Logging
python-json-logger>=2.0.7

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0

# Utilities
schedule>=1.2.0
pyyaml>=6.0.1

# Terminal UI
rich>=13.7.0

# Web Dashboard
streamlit>=1.31.0

# Statistical Analysis
scipy>=1.11.4

# Type Checking (Development)
mypy>=1.7.1
types-requests>=2.31.0.10
types-pyyaml>=6.0.12.12

# Code Quality (Development)
pylint>=3.0.3
black>=23.12.1
flake8>=6.1.0
